Change the factory class cast and refactor test
